feat: create tray icon service
